+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/). This project follows the conventions of [semver.org](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added the option to pass in a fixed list of tables from the command line rather than scanning the full database.
+- Added a command line option (`--ignore` / `-i`) to ignore a column during data generation.
+- Added a null data generation option to give the user a way to force a column to null.
+
 ### Changed
 - Changed the long option for the output directory to `--directory` for `--output-dir`.
-- Added the option to pass in a fixed list of tables from the command line rather than scanning the full database.
 
 ## [0.4.1] - 2017-08-25
 ### Fixed